refactor(racingcargame): 입출력 명확하게 나누기
- 역할에 따라 알맞는 이름으로 변경
- 컨트롤러에서 만들어진 기능들을 단순 제어하게 수정
