+++ src/main/java/com/nextstep/racingcar/views/OutputView.java
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar.view;
+package com.nextstep.racingcar.views;
 
 import com.nextstep.racingcar.domain.Car;
 import java.util.Collections;
@@ -9,16 +9,8 @@ public class OutputView {
 
   private final String CAR_UNIT_STEP_SYMBOL = "-";
 
-  public void printRequestCarNames() {
-    printer("경주할 자동차 이름을 입력하세요(이름은 쉽표(,)를 기준으로 구분).");
-  }
-
-  public void printRequestRepeatCount() {
-    printer("시도할 회수는 몇회인가요?");
-  }
-
   public void printResultGuide() {
-    printer("실행 결과");
+    System.out.println("실행 결과");
   }
 
   public void printSimulationResults(List<Car> cars) {
@@ -26,7 +18,7 @@ public class OutputView {
       printCarStatus(car);
     }
 
-    printer("");
+    System.out.println("");
   }
 
   public void printWinners(List<Car> winners) {
@@ -34,7 +26,7 @@ public class OutputView {
         .map(Car::getName)
         .collect(Collectors.joining(", "));
 
-    printer(String.format("%s가 최종 우승했습니다.", winnersAppended));
+    System.out.println(String.format("%s가 최종 우승했습니다.", winnersAppended));
   }
 
   private void printCarStatus(Car car) {
@@ -45,10 +37,6 @@ public class OutputView {
             Collections.nCopies(car.getLocation(), CAR_UNIT_STEP_SYMBOL)
         ));
 
-    printer(status);
-  }
-
-  private void printer(String message) {
-    System.out.println(message);
+    System.out.println(status);
   }
 }
